refactored the manga data struct, all info shows up for mangas that are searched, i now want to see if i can search by tags.
